Split examples into separate files
